workaround for lein parallel-test
lein parallel-test fails, because the plugin loads the old version of
core.async on start, so otplike can not be compiled. The workaround is
to tell parallel-test the right version of core.async as a part of
special :parallel-test profile.
Unfortunately, parallel-test has a bug preventing custom configuration
from being used. Thus the patched version of the plugin is required at
the moment.