process: rework process function
Process function is instance of TProcFn now and not a function. It can
only be spawned, not applied. Recursion "by name" is not allowed.
Trace events and process info contain initial call as a symbol (not
function).
`proc-fn` accepts the name of a process function as a way to spawn it
again from the inside.